[fix] merge MongoDB 5/6 testcases
