--- flink-connector-mongodb-cdc/src/test/java/com/ververica/cdc/connectors/mongodb/mongo6/MongoDB6Container.java
@@ -1,126 +1,0 @@
-/*
- * Copyright 2022 Ververica Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ververica.cdc.connectors.mongodb.mongo6;
-
-import com.ververica.cdc.connectors.mongodb.utils.MongoDBContainer;
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.output.OutputFrame;
-
-import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Random;
-import java.util.function.Consumer;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import static org.junit.Assert.assertNotNull;
-
-/** Container for testing MongoDB >= 5.0.3. */
-public class MongoDB6Container extends org.testcontainers.containers.MongoDBContainer {
-
-    private static final Logger LOG = LoggerFactory.getLogger(MongoDB6Container.class);
-
-    private static final Pattern COMMENT_PATTERN = Pattern.compile("^(.*)//.*$");
-
-    public static final int MONGODB_PORT = 27017;
-
-    MongoDB6Container(String imageName) {
-        super(imageName);
-    }
-
-    @Override
-    public MongoDB6Container withSharding() {
-        return (MongoDB6Container) super.withSharding();
-    }
-
-    @Override
-    public MongoDB6Container withLogConsumer(Consumer<OutputFrame> consumer) {
-        return (MongoDB6Container) super.withLogConsumer(consumer);
-    }
-
-    public void executeCommand(String command) {
-        try {
-            LOG.info("Executing mongo command: {}", command);
-            ExecResult execResult = execInContainer("mongosh", "--eval", command);
-            LOG.info(execResult.getStdout());
-            if (execResult.getExitCode() != 0) {
-                throw new IllegalStateException(
-                        "Execute mongo command failed " + execResult.getStdout());
-            }
-        } catch (InterruptedException | IOException e) {
-            throw new IllegalStateException("Execute mongo command failed", e);
-        }
-    }
-
-    public String executeCommandInDatabase(String command, String databaseName) {
-        try {
-            executeCommand(String.format("db = db.getSiblingDB('%s');\n", databaseName) + command);
-            return databaseName;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    /** Executes a mongo command file in separate database. */
-    public String executeCommandFileInSeparateDatabase(String fileNameIgnoreSuffix) {
-        return executeCommandFileInDatabase(
-                fileNameIgnoreSuffix,
-                fileNameIgnoreSuffix + "_" + Integer.toUnsignedString(new Random().nextInt(), 36));
-    }
-
-    /** Executes a mongo command file, specify a database name. */
-    public String executeCommandFileInDatabase(String fileNameIgnoreSuffix, String databaseName) {
-        final String dbName = databaseName != null ? databaseName : fileNameIgnoreSuffix;
-        final String ddlFile = String.format("ddl/%s.js", fileNameIgnoreSuffix);
-        final URL ddlTestFile = MongoDBContainer.class.getClassLoader().getResource(ddlFile);
-        assertNotNull("Cannot locate " + ddlFile, ddlTestFile);
-
-        try {
-            // use database;
-            String command0 = String.format("db = db.getSiblingDB('%s');\n", dbName);
-            String command1 =
-                    Files.readAllLines(Paths.get(ddlTestFile.toURI())).stream()
-                            .filter(x -> StringUtils.isNotBlank(x) && !x.trim().startsWith("//"))
-                            .map(
-                                    x -> {
-                                        final Matcher m = COMMENT_PATTERN.matcher(x);
-                                        return m.matches() ? m.group(1) : x;
-                                    })
-                            .collect(Collectors.joining("\n"));
-
-            executeCommand(command0 + command1);
-
-            return dbName;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public String getConnectionString() {
-        return String.format(
-                "mongodb://%s:%d", getContainerIpAddress(), getMappedPort(MONGODB_PORT));
-    }
-
-    public String getHostAndPort() {
-        return String.format("%s:%s", getContainerIpAddress(), getMappedPort(MONGODB_PORT));
-    }
-}
