--- flink-connector-mongodb-cdc/src/test/java/com/ververica/cdc/connectors/mongodb/mongo6/MongoDB6SourceTestBase.java
@@ -1,76 +1,0 @@
-/*
- * Copyright 2022 Ververica Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ververica.cdc.connectors.mongodb.mongo6;
-
-import org.apache.flink.runtime.minicluster.RpcServiceSharing;
-import org.apache.flink.runtime.testutils.MiniClusterResourceConfiguration;
-import org.apache.flink.test.util.MiniClusterWithClientResource;
-
-import com.mongodb.ConnectionString;
-import com.mongodb.MongoClientSettings;
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.lifecycle.Startables;
-
-import java.util.stream.Stream;
-
-/** MongoDBSourceTestBase for MongoDB >= 5.0.3. */
-public class MongoDB6SourceTestBase {
-
-    protected static MongoClient mongodbClient;
-
-    protected static final int DEFAULT_PARALLELISM = 4;
-
-    @Rule
-    public final MiniClusterWithClientResource miniClusterResource =
-            new MiniClusterWithClientResource(
-                    new MiniClusterResourceConfiguration.Builder()
-                            .setNumberTaskManagers(1)
-                            .setNumberSlotsPerTaskManager(DEFAULT_PARALLELISM)
-                            .setRpcServiceSharing(RpcServiceSharing.DEDICATED)
-                            .withHaLeadershipControl()
-                            .build());
-
-    @BeforeClass
-    public static void startContainers() {
-        LOG.info("Starting containers...");
-        Startables.deepStart(Stream.of(CONTAINER)).join();
-
-        MongoClientSettings settings =
-                MongoClientSettings.builder()
-                        .applyConnectionString(
-                                new ConnectionString(CONTAINER.getConnectionString()))
-                        .build();
-        mongodbClient = MongoClients.create(settings);
-
-        LOG.info("Containers are started.");
-    }
-
-    private static final Logger LOG = LoggerFactory.getLogger(MongoDB6SourceTestBase.class);
-
-    @ClassRule
-    public static final MongoDB6Container CONTAINER =
-            new MongoDB6Container("mongo:6.0.6")
-                    .withSharding()
-                    .withLogConsumer(new Slf4jLogConsumer(LOG));
-}
