+++ flink-connector-mongodb-cdc/src/test/java/com/ververica/cdc/connectors/mongodb/source/reader/MongoDBStreamSplitReaderTest.java
@@ -83,11 +83,11 @@ public class MongoDBStreamSplitReaderTest extends MongoDBSourceTestBase {
 
     @Before
     public void before() {
-        database = ROUTER.executeCommandFileInSeparateDatabase("chunk_test");
+        database = CONTAINER.executeCommandFileInSeparateDatabase("chunk_test");
 
         MongoDBSourceConfigFactory configFactory =
                 new MongoDBSourceConfigFactory()
-                        .hosts(ROUTER.getHostAndPort())
+                        .hosts(CONTAINER.getHostAndPort())
                         .databaseList(database)
                         .collectionList(database + ".shopping_cart")
                         .username(FLINK_USER)
