+++ flink-connector-mongodb-cdc/src/test/java/com/ververica/cdc/connectors/mongodb/table/MongoDBRegexFilterITCase.java
@@ -76,9 +76,9 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
     public void testMatchMultipleDatabasesAndCollections() throws Exception {
         // 1. Given collections:
         // db0: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db0 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db0 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
         // db1: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db1 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db1 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
 
         // 2. Test match: collection = ^(db0|db1)\.coll_a\d?$
         String collectionRegex = String.format("^(%s|%s)\\.coll_a\\d?$", db0, db1);
@@ -119,11 +119,11 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
     public void testMatchMultipleDatabases() throws Exception {
         // 1. Given collections:
         // db0: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db0 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db0 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
         // db1: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db1 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db1 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
         // db2: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db2 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db2 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
 
         // 2. Test match database: ^(db0|db1)$
         String databaseRegex = String.format("%s|%s", db0, db1);
@@ -173,9 +173,9 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
     public void testMatchSingleQualifiedCollectionPattern() throws Exception {
         // 1. Given collections:
         // db0: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db0 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db0 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
         // db1: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db1 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db1 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
 
         // 2. Test match: collection ^(db0|db1)\.coll_a\d?$
         String collectionRegex = String.format("^%s\\.coll_b\\d?$", db0);
@@ -212,9 +212,9 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
     public void testMatchSingleDatabaseWithCollectionPattern() throws Exception {
         // 1. Given collections:
         // db0: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db0 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db0 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
         // db1: [coll_a1, coll_a2, coll_b1, coll_b2]
-        String db1 = ROUTER.executeCommandFileInSeparateDatabase("ns_regex");
+        String db1 = CONTAINER.executeCommandFileInSeparateDatabase("ns_regex");
 
         // 2. Test match: collection .*coll_b\d?
         String collectionRegex = ".*coll_b\\d?";
@@ -250,7 +250,7 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
     public void testMatchDatabaseAndCollectionContainsDash() throws Exception {
         // 1. Given collections:
         // db0: [coll-a1, coll-a2, coll-b1, coll-b2]
-        String db0 = ROUTER.executeCommandFileInSeparateDatabase("ns-regex");
+        String db0 = CONTAINER.executeCommandFileInSeparateDatabase("ns-regex");
 
         TableResult result = submitTestCase(db0, "coll-a1");
 
@@ -275,7 +275,7 @@ public class MongoDBRegexFilterITCase extends MongoDBSourceTestBase {
                         + " coll_name STRING METADATA FROM 'collection_name' VIRTUAL,"
                         + " PRIMARY KEY (_id) NOT ENFORCED"
                         + ") WITH ("
-                        + ignoreIfNull("hosts", ROUTER.getHostAndPort())
+                        + ignoreIfNull("hosts", CONTAINER.getHostAndPort())
                         + ignoreIfNull("username", FLINK_USER)
                         + ignoreIfNull("password", FLINK_USER_PASSWORD)
                         + ignoreIfNull("database", database)
