+++ flink-connector-mongodb-cdc/src/test/resources/docker/mongodb/setup.js
@@ -42,3 +42,5 @@ db.createUser(
    ]
  }
 );
+
+console.log('Flink test user created.\n');
